made result id field an int
Former-commit-id: 2cf08abeb314b28ee66ece373c8f586fb6984082
Former-commit-id: dc8353644b60f378a1de8ee2ef009c440da49d4c

+++ src/mitll/langtest/server/LangTestDatabaseImpl.java
@@ -1244,7 +1244,7 @@ public class LangTestDatabaseImpl extends RemoteServiceServlet implements LangTe
   }
 
   @Override
-  public void addRoundTrip(long resultID, int roundTrip) {
+  public void addRoundTrip(int resultID, int roundTrip) {
     db.getAnswerDAO().addRoundTrip(resultID, roundTrip);
   }
